refactor(extension/podman): mv darwin test to dedicated file
